requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
python-multipart==0.0.9
pillow
ultralytics==8.1.0
python-dotenv
torch==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.20.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu